requirements.txt
Flask==2.3.3
monero==1.1.1
requests==2.31.0
python-dotenv==1.0.0
